backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
joblib==1.5.2
scikit-learn==1.7.2
xgboost==3.1.2
pandas==2.3.3
numpy==2.3.5

